refactor: move server/infra tests into test/server/infra
